backadmin/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.5.0
djongo==1.3.6
dnspython==1.16.0
pika == 1.2.0